+++ inst/resources/init-rprofile.R
@@ -1,3 +1,3 @@
-#### -- Packrat Autoloader (version 0.4.1.16) -- ####
+#### -- Packrat Autoloader (version 0.4.1.17) -- ####
 source("packrat/init.R")
 #### -- End Packrat Autoloader -- ####